DRY up request logic into request file.
