+++ lib/Issuer.js
@@ -1,14 +1,7 @@
 /*!
  * Copyright (c) 2022 Digital Bazaar, Inc. All rights reserved.
  */
-import {httpClient} from '@digitalbazaar/http-client/main.js';
-import {ZcapClient} from '@digitalbazaar/ezcap/lib/main.js';
-import {decodeSecretKeySeed} from 'bnid/main.js';
-import {
-  Ed25519Signature2020
-} from '@digitalbazaar/ed25519-signature-2020/lib/main.js';
-import {constructOAuthHeader} from './oath2';
-import {agent, didKeyDriver} from './constants.js';
+import {httpRequest, zcapRequest} from './requests.js';
 
 export class Issuer {
   constructor({oath2, issuer}) {
@@ -29,99 +22,36 @@ export class Issuer {
    */
   async issue({body}) {
     const {issuer, oath2} = this;
+    const headers = {...issuer.headers};
     if(issuer.zcap) {
-      return _zcapClientRequest({
-        ...issuer,
+      return zcapRequest({
+        endpoint: issuer.endpoint,
+        zcap: issuer.zcap,
         json: body
       });
     }
-    return _httpRequest({
-      ...issuer,
+    return httpRequest({
+      endpoint: issuer.endpoint,
       oath2,
-      json: body
+      json: body,
+      headers
     });
   }
   async setStatus({body}) {
-    const headers = {
-      ...this.settings.issuer.headers
-    };
-    try {
-      let result;
-      if(this.settings.issuer.zcap) {
-        result = await _zcapClientRequest({
-          endpoint: this.settings.issuer.statusEndpoint,
-          zcap: this.settings.zcap,
-          json: body
-        });
-      } else {
-        result = await httpClient.post(
-          this.settings.issuer.statusEndpoint,
-          {headers, agent, json: body});
-      }
-      return result;
-    } catch(e) {
-      console.log(e);
-      throw e;
-    }
-  }
-}
-
-async function _httpRequest({endpoint, json, oath2, headers = {}}) {
-  let result;
-  let error;
-  if(oath2) {
-    headers.Authorization = await constructOAuthHeader({...oath2});
-  }
-  try {
-    result = await httpClient.post(
-      endpoint,
-      {
-        json,
-        agent,
-        headers: {...headers}
+    const {issuer, oath2} = this;
+    const headers = {...issuer.headers};
+    if(issuer.zcap) {
+      return zcapRequest({
+        endpoint: issuer.statusEndpoint,
+        zcap: this.settings.zcap,
+        json: body
       });
-  } catch(e) {
-    // delete the Authorization header to prevent
-    // oath2 headers in reports
-    if(e.response) {
-      e.response.headers.delete('Authorization');
     }
-    error = e;
-  }
-  return {result, error};
-}
-
-const _getZcapClient = async ({secretKeySeed}) => {
-  const seed = await decodeSecretKeySeed({secretKeySeed});
-  const didKey = await didKeyDriver.generate({seed});
-  const {didDocument: {capabilityInvocation}} = didKey;
-  return new ZcapClient({
-    SuiteClass: Ed25519Signature2020,
-    invocationSigner: didKey.keyPairs.get(capabilityInvocation[0]).signer(),
-    agent
-  });
-};
-
-async function _zcapClientRequest({endpoint, json, zcap, headers = {}}) {
-  let result;
-  let error;
-  let capability = zcap.capability;
-  // we are storing the zcaps stringified right now
-  if(typeof zcap.capability === 'string') {
-    capability = JSON.parse(capability);
-  }
-  try {
-    // assume that the clientSecret is set in the test environment
-    const secretKeySeed = process.env[zcap.clientSecret];
-    const zcapClient = await _getZcapClient({secretKeySeed});
-    result = await zcapClient.write({
-      url: endpoint,
-      json,
-      headers: {...headers},
-      capability
+    return httpRequest({
+      json: body,
+      oath2,
+      endpoint: issuer.statusEndpoint,
+      headers
     });
-  } catch(e) {
-    error = e;
   }
-  return {result, error};
 }
