+++ lib/Verifier.js
@@ -3,49 +3,30 @@
  */
 'use strict';
 
-import {httpClient} from '@digitalbazaar/http-client/main.js';
-import {agent} from './constants.js';
+import {httpRequest, zcapRequest} from './requests.js';
 
 export class Verifier {
   constructor({verifier, oath2}) {
     this.verifier = verifier;
     this.oauth2 = oath2;
   }
-  async verify({credential, auth}) {
-    try {
-      const body = {
-        verifiableCredential: credential,
-        options: {
-          checks: ['proof', 'credentialStatus'],
-        },
-      };
-      const headers = {
-        ...this.settings.headers
-      };
-      if(auth && auth.type === 'oauth2-bearer-token') {
-        headers.Authorization = `Bearer ${auth.accessToken}`;
-      }
-      let result;
-      if(this.verifier.zcap) {
-        const zcapClient = await _getZcapClient();
-        result = await zcapClient.write({
-          url: this.verifier.endpoint,
-          capability: JSON.parse(this.settings.verifier.zcap),
-          json: body
-        });
-      } else {
-        result = await httpClient.post(
-          this.settings.endpoint,
-          {headers, agent, json: body}
-        );
-      }
-      return result;
-    } catch(e) {
-      // this is just to make debugging easier
-      if(e && e.response && e.response.data) {
-        throw new Error(JSON.stringify(e.response.data, null, 2));
-      }
-      throw e;
+  async verify({body}) {
+    const {verifier, oath2} = this;
+    const headers = {
+      ...verifier.headers
+    };
+    if(verifier.zcap) {
+      return zcapRequest({
+        endpoint: verifier.endpoint,
+        zcap: verifier.zcap,
+        json: body
+      });
     }
+    return httpRequest({
+      json: body,
+      oath2,
+      endpoint: verifier.endpoint,
+      headers
+    });
   }
 }
